+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 All notable changes to thsi project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/).
+The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+### Added
+- '-p' flag to prompt on all task changing commands
 
 ## [0.4.0] - 2018-04-18
 ### Changed