+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.1] - 2018-04-29
 ### Added
 - '-p' flag to prompt on all task changing commands
 - Added "addx" command to add completed tasks
+
+### Changed
+- Contexts support GTD "@@" and "@\_" names
 
 ## [0.4.0] - 2018-04-18
 ### Changed